Move KonashiUtils to util package
