+++ konashi-android-sdk/src/main/java/com/uxxu/konashi/lib/util/KonashiUtils.java
@@ -1,4 +1,4 @@
-package com.uxxu.konashi.lib;
+package com.uxxu.konashi.lib.util;
 
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.util.Log;
